+++ test/bigtest/network/serializers/item.js
@@ -1,0 +1,3 @@
+import { buildBaseSerializer } from './base';
+
+export default buildBaseSerializer('items', 'items');