+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 ### Stories
+* [UIOR-444](https://issues.folio.org/browse/UIOR-444) common constants for Piece, BigTest config
 * Update to stripes v4
 * [UIREC-42](https://issues.folio.org/browse/UIREC-42) Filter Titles by piece status and Acq unit
 * [UIF-202](https://issues.folio.org/browse/UIF-202) Add default sorting param to AcqList